move slack blueprint to slack integration app
